feat(strawberry-poc): close client-divergence gaps for weka, runzi, nzshm-model
Wire-format fixes driven by surveying real client queries across weka
(UI), nzshm-runzi (Python automation), and nzshm-model (Python library):

- Rename POC SDL type ToshiFile -> File (Python class unchanged).
  weka and nzshm-model query `... on File`; ToshiFile was POC-only.
- Rename NodeFilterPayload SDL type to NodeFilter.
  weka Relay codegen embeds `concreteType: "NodeFilter"` literally.
- Add RowItemType enum and use it on Table.column_types and
  CreateTableInput.column_types. runzi mutations declare
  `$column_types: [RowItemType]!` — without the enum, table creation
  from runzi was dead against the POC.
- Add `inversion_solution: InversionSolutionUnion` resolver on
  AutomationTask, RuptureGenerationTask, and OpenquakeHazardTask.
  weka's GeneralTask children pages traverse this field. Resolver
  traverses files_raw and picks the WRITE-related InversionSolution
  variant, matching legacy semantics.

Tests: 161 pass (151 + 10 new in test_weka_parity.py).

Cross-PR consolidation tweaks:
- test_file_relation_compression: `... on ToshiFile` → `... on File`
  (the test was added by #299 before the SDL rename in #309)
- test_client_mutation_id_compat: lower payload-count threshold
  (Phase 2 renamed several Create*Payload → Create*; expected count drops)

+++ spike/strawberry_poc/tests/test_inversion_solution_interface.py
@@ -112,7 +112,7 @@ def _seed_mfd_table(gql_context, rgt_id: str) -> str:
                 "table_type": "MFD_CURVES_V2",
                 "created": "2024-01-01T00:00:00Z",
                 "column_headers": ["mag", "rate"],
-                "column_types": ["float", "float"],
+                "column_types": ["double", "double"],
                 "rows": [["7.0", "1e-5"], ["7.5", "5e-6"]],
             }
         },
